UI changing. Added signup page.

+++ lab9/nikulin/src/main/java/edu/karazin/shop/model/OrderList.java
@@ -18,7 +18,7 @@ public class OrderList {
     private int userId;
 
     @Column(name = "total_price")
-    private Integer totalPrice;
+    private Double totalPrice;
 
     @ManyToMany(cascade = CascadeType.DETACH)
     @LazyCollection(LazyCollectionOption.FALSE)
@@ -31,7 +31,6 @@ public class OrderList {
     private List<BookList> bookLists;
 
     @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
-//    @LazyCollection(LazyCollectionOption.FALSE)
     @JoinTable(name = "order_order_details",
             joinColumns = {
                     @JoinColumn(name = "order_id")
@@ -72,11 +71,11 @@ public class OrderList {
         this.orderDetails = orderDetails;
     }
 
-    public Integer getTotalPrice() {
+    public Double getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(Double totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
